Refactoring to share table panel between all candidates pages
