+++ static/js/pages/candidates.js
@@ -1,13 +1,10 @@
 'use strict';
 
 var $ = require('jquery');
-var _ = require('underscore');
 
 var tables = require('../modules/tables');
-var helpers = require('../modules/helpers');
 var columns = require('../modules/columns');
-
-var candidatesTemplate = require('../../templates/candidates.hbs');
+var tablePanels = require('../modules/table-panels');
 
 $(document).ready(function() {
   var $table = $('#results');
@@ -19,7 +16,7 @@ $(document).ready(function() {
     useExport: true,
     rowCallback: tables.modalRenderRow,
     callbacks: {
-      afterRender: tables.modalRenderFactory(candidatesTemplate)
+      afterRender: tablePanels.renderCandidatePanel(false)
     }
   });
 });
