+++ static/templates/candidates.hbs
@@ -1,29 +1,99 @@
 <div class="panel__row">
-  <div class="panel__heading">
-    <h4 class="panel__title"><a href="{{basePath}}/candidate/{{candidate_id}}">{{name}}</a></h4>
-  </div>
+  <h4 class="panel__title">
+    <a href="{{entityUrl this type="candidate" query=query}}">{{name}}</a>
+  </h4>
+</div>
+<div class="panel__row">
+  <h4 class="panel__title">Candidate information</h4>
   <table>
-    {{#panelRow "Office"}}
-      {{office_full}}
+    {{#panelRow "Candidate ID"}}
+      {{candidate_id}}
     {{/panelRow}}
-    {{#panelRow "Election year"}}
-      {{election_years}}
+    <tr>
+      <td class="panel__term">Election years</td>
+      <td class="panel__data">
+        {{#each election_years}}
+          <a href="{{electionUrl this parentContext=../this index=@index}}">{{this}}</a>{{#unless @last}}, {{/unless}}
+        {{/each}}
+        </ul>
+      </td>
+    </tr>
+    {{#panelRow "Office sought"}}
+      {{office_full}}
     {{/panelRow}}
     {{#panelRow "Political party"}}
       {{party_full}}
     {{/panelRow}}
-    {{#panelRow "Status"}}
-      {{incumbent_challenge_full}}
-    {{/panelRow}}
-    {{#if state}}
+    {{#unless (eq office 'P')}}
       {{#panelRow "State"}}
-        {{state}}
+        {{decodeState state}}
       {{/panelRow}}
-    {{/if}}
-    {{#if district}}
+    {{/unless}}
+    {{#if (eq office 'H')}}
       {{#panelRow "District"}}
         {{district}}
       {{/panelRow}}
     {{/if}}
+    {{#if committees.principal}}
+      {{#panelRow "Principal campaign committees"}}
+      <ul>
+        {{#each committees.principal}}
+          <li>
+            <a href="{{entityUrl this type="committee"}}">
+              {{ this.name}}
+            </a>
+          </li>
+        {{/each}}
+      </ul>
+      {{/panelRow}}
+    {{/if}}
+    {{#if committees.authorized}}
+      {{#panelRow "Authorized campaign committees"}}
+        {{#each committees.authorized}}
+          <li>
+            <a href="{{entityUrl this type="committee"}}">
+              {{ this.name}}
+            </a>
+          </li>
+        {{/each}}
+      {{/panelRow}}
+    {{/if}}
+    {{#if first_form_2}}
+      {{#panelRow "First statement of candidacy"}}
+        <a href="{{first_form_2.pdf_url}}">{{first_form_2.document_description}}</a>
+        <span class="t-block">Filed: {{datetime first_form_2.receipt_date format='pretty'}}</span>
+      {{/panelRow}}
+    {{/if}}
+    {{#if last_form_2}}
+      {{#panelRow "Most recent statement of candidacy"}}
+        <a href="{{last_form_2.pdf_url}}">{{last_form_2.document_description}}</a>
+        <span class="t-block">Filed: {{datetime last_form_2.receipt_date format='pretty'}}</span>
+      {{/panelRow}}
+    {{/if}}
+
   </table>
 </div>
+{{#if showFinancialTotals}}
+<div class="panel__row">
+  <h4 class="panel__title">Financial totals for {{time_period}}</h4>
+  <table>
+    {{#panelRow "Receipts"}}
+      {{{currency receipts}}}
+    {{/panelRow}}
+    {{#panelRow "Disbursements"}}
+      {{{currency disbursements}}}
+    {{/panelRow}}
+    {{#panelRow "Ending cash on hand"}}
+      {{{currency cash_on_hand_end_period}}}
+    {{/panelRow}}
+    {{#panelRow "Debts owed by committee"}}
+      {{{currency debts_owed_by_committee}}}
+    {{/panelRow}}
+    {{#if (eq office 'P')}}
+      {{#panelRow "Has accepted presidential public funds"}}
+        {{federal_funds_flag}}
+      {{/panelRow}}
+    {{/if}}
+  </table>
+</div>
+{{/if}}
